Fix rust_scorer build: widen synapse from_index u16→u32 for GPU
neat-core narrowed `Synapse::from_index` from u32 to u16 (neat-core #177).
The GPU SSBO mirror `SynapseGpu.from_index` and the WGSL shaders still
require u32, so building `build_batched_network_data` failed with E0308
(expected u32, found u16) at forward_mse_batched.rs:228.

Widen with `u32::from(s.from_index)`, matching the existing conversions
for `num_synapses`/`squash_type` a few lines above. Verified with
`cargo build --release -p rust_scorer` against neat-core Develop.

Vibe-Coder-Run-Id: vibe-mqqg2mpj-c37c8f

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core narrowed `from_index` to u16 (neat-core #177); the
+                // GPU SSBO layout and WGSL shaders require u32, so widen here.
+                from_index: u32::from(s.from_index),
             });
         }
